Extract GCP mocks into separate class
